Implement generate-apply-based utilities
For implementing generate-apply based controllers, it's often important
to know the objects that are owned by a controller.

For doing so, the utilities provided with this change allow to
* Filter owned objects
* List and filter
* List and filter owned objects
* Get objects from
  * Slices
  * Lists
* Set objects to
  * Slices
  * Lists

And, most prominently, the new beta-grade tool called
`CreateOrUse` that tries to use or create an object depending on a
predefined list.